refact: move Snap to separate file
